Trunk: Merged EXP-port at r18975 to the trunk.
This change is the result of EMC's initial changes for improved porting of the code to Zeus.  The changes are intended to make running on Zeus possible and the transition to other machines more straightforward.  Note the these changes, also include bugs or inconsistencies in the code that were found in the process.  Also note, that these changes will probably not be the final changes.  These changes did not include the final fixes for all OPENMP problems, additional optimizations are possible and has not been tested on Zeus for all regression tests (they have been tested on the IBM for all regression tests), since not all regression tests have been ported. The update includes regression test modifications to make the GDAS and NMMB regression tests work on Zeus.

The regression tests have been performed on this version and all results verified.  The only changes result from fixing a bug in the combine_radobs routine which would have allowed different results using different sets of observations when the criterion is the same on multiple processors.  The fix removes the possibility of different results, but may change results when this situation occurs.  This does change results in several of the EMC regression tests.  Also, moving the 0.001 multiplication from ratio_errors to the ramp variable if l_sfcobserror_ramp_q is true.  This will also affect results for all EMC regression tests.

Please read ticket #247 for details