gmted2010 branch:
Add back the bacio library to the build process.  Without it,
the compile will fail for "-O0".  When using "-O3", the
compile with complete without the bacio library.  Not sure why this is.
Also, map in the mixed precision version of all libraries.
Some builds were using the double precision version.

Ticket #17.