cycle_fv3 branch:
Several updates to chgres.f90 - In routine "adjust_nsst"
compute the maximum search to an open water GSI point based
on the resolution of the GSI data.  Previously the max
search was hardwired.  Add max/min range checking for
TREF and skin temp.  New routine "climo_trend" that 
updates TREF based on the annual cycle of the RGTSST
climatology.  This update is only used when the search
fails to find an open water GSI point within the max
search radius.

Ticket #20.