cycle_fv3 branch:
The interpolation of the GSI TREF increment to the model native
grid is now bilinear to be consistent with how the forecast
model interpolates atmospheric increments.  The model routine
that computes the bilinear weights - remap_coef - was incorporated
in cycle.f90.  This routine assumes the SP is row '1', so 
remove the 'flip' of the GSI data in routine "read_write_data".

Ticket #20.