port2dell branch:  This commit references #49439.
CHGRES - Update logic that determines whether the input
atmospheric restart file is "nemsio" or "sigio" format.
Previously, the sigio library attempted to read the file
header.  If the file was "nemsio", a bad status was passed
back from the library.  On Dell, the sigio library will
seg fault when trying to read a nemsio file.  The seg fault
results from compiling the sigio library with Intel v18,
which is the only option available on Dell.  The new
logic checks for the string "NEMSIO" in the first 16 bytes
of the atmospheric file.  If the string exists, it is
"nemsio".  Otherwise, it is assumed to be "sigio".

	modified:   sorc/global_chgres.fd/chgres.f90

Change-Id: I21785a3c73b5e4ba31320b17b5d9adb90b0dc982