reduce memory usage of enkf_chgres_recenter by:
1. Interpolating horizontally during the read of the input file.  This eliminates the need to store the input grid in memory; only 2 2D slices are needed.

2. Interpolate vertically one I-K slice at a time.  This reduces the vertical interp part from having five output grids in memory to only two.

At the operational resolution, this reduces the memory usage from ~40.5 GB down to ~15.5 GB.

Without jumping through hoops of fire, it is not possible to reduce memory usage significantly unless the program is MPI parallelized.  These changes are also designed to make it easier to MPI parallelize, once that need is established.