Add script to combine FV3 and chem tracer tiles
In order to run the FV3 model with GSDCHEM from a cold start, we need to
merge chemistry tracer data into the FV3 atmosphere initial condition.
This commit adds a script that combines the chemistry tracers into the
atmospheric file and updates the tracer dimension for a single tile.

A module adding python3 to the path must be loaded prior to running this
script.

The script takes three arguments: the atmosphere file, the chemistry
file, and a text file with a list of the tracers to add from the
chemistry file. An optional fourth argument is available if you wish to
output to a new file instead of overwriting the atmosphere file. Full
usage details are available  by printing the help message with -h or
--help:

    merge_fv3_chem_tile.py --help

Change-Id: I66ccaf1c8b7906d5906f2326fe0ad699e3ed4ee6
Refs: #62104