Update NEMS app version to a master commit
Previously the checkout for the FV3-GSDCHEM-WW3 app temporarily pointed to an
interim commit that included GSDCHEM 0.87 and used ESMF 8.0.0bs40+ to facilitate
development while waiting for the app to merge a full update to master. That
update has now been made to the app master, so the checkout commit is now updated
to point to that new version.

The new FV3-GSDCHEM-WW3 app commit updates to GSDCHEM 0.87 and ESMF 8.0.0bs47,
and also updates the other components to recent development versions.

Change-Id: I0318535aa7564b48567a6442cdda074007095c03
Refs: #58418