Add chemistry coupling with GSDCHEM 0.88
Changes Lin made outside of version control to run with aerosols is
combined with existing script that allows wave coupling. Right now
the two cannot be combined at the same time. That functionality will
need to be updated at a future time.

For consistency with cplwav, chemistry coupling is turned on with
cplchm. The prep directory is assumed to be in the INPUT directory.

Some settings used for testing have been retained but commented out.
They will be removed in a future commit.

Change-Id: I40a647ca522121c2c3eb1ceed45c45c8c8547f44
Refs: #62104