Change way wave restart files are produced
In order to facilitate faster cycle turnover, instead of copying
the wave restart files at the end of the forecast, symbolic links
to the final destination are created before the forecast. This
allows the restart directory to be populated immediately, so that
the next cycle can commence while the current cycle finishes.

Change-Id: I887dddec88e32ad08fefc5a8080c9f553965b9dc