Update exglobal_fcst to reduce unnecessary links for GEFS
Some modifications to reduce the depth of symlinks for GEFS. The
ICSDIR, which had been defined but unused, is now used as the source
of cold-start initial conditions. The default is $memdir/INPUT, so
the source is unchanged if not set. Additionally, a FCSTDIR variable
is added to allow an output location other than $memdir. The default
if not set is $memdir, so functionality is unchanged if not set.

The GEFS workflow was changed to use a CDUMP of the ensemble member
instead of gdas as previous. A new setting (gefs) detects this by
looking at the first two characters of CDUMP, and that switch is used
throughout were needed to select the correct branch.

Additionally, some namelist settings needed for restart of stochastic
patterns are added.