exglobal_fcst fix for non-divisible restart intervals
Because exglobal_fcst looked backwards from the end of the forecast
for restart files, it would never pick up restart files if the
forecast length was not evenly divisible by the restart interval.
Now the modulus is subtracted to the end time first, so that the
correct restart times are checked.