Move utils linking out of build script
The linker script for ufs_utils was being called from the build
script instead of the master link script and was hard-coded to
use emc (symlink). The call is now moved to the master linker
script and properly uses the command line argument to determine
whether to copy (nco) or use symlinks (emc).