Update build for WCOSS Dell
The build for the ww3 and reg2grb2 programs are updated to build on
WCOSS Dell. New modules are created for each. The two modules are both
using development modules; this needs to be updated to NCO modules in
the future. ww3 uses the newer versions of modules under NCEPLIBS-v1.2.0,
but the reg2grb2 was unable to use that location due to an issue with
wgrib2/2.0.8. reg2grb2 is also using older modules than ww3 on Hera, so
this mismatch already exists.

The build script for reg2grb2 is updated to use the environment
variables set by the wgrib2 module on WCOSS. It should remain backwards
compatible for Hera/Orion, but this needs to be tested.

Refs: #175