Fix build of WW3 pre/post on WCOSS
The build of WW3 pre/post was returning as successful despite failing
to compile some of the executables. The issue was traced back to some
of the library paths not being set correctly. The build script is
updated to use the existing environment variables for those paths if
they are already set, as they are by the modules on WCOSS.

Refs: #175