Add missing defaults for chem settings
Due to the way CROW is currently set up, all of the chem_settings
need to be defined even if they aren't used or relevant. Defaults
for the missing settings have been added by using the string
"not_set" when cplgocart is not true.

This arrangment results in a slightly different error message if
a case with gocart on fails to set one of the necessary settings.
Instead of getting a missing variable exception:

crow.config.exceptions.VariableMissing:
  doc.chem_settings[0]: missing:
    CFGDIR, INPDIR in: CDUMP, Template, MOD, ntdiag

You get a type exception that the type is null with the misleading
suggestion that it should be in a !calc:

doc.chem_settings[0].CFGDIR=null:
  not valid for type 'string'.  Should this have been a !calc?

Refs: #304