Clean up and add new settings in metp.sh
SDATE_GFS was added into the script preamble following the
addition of the variable being added as a rocoto exported
environment variable for metp tasks.

Config.vrfy was removed from the list of relevant configs
to source.

Variables CDATEm1 and PDYm1 were removed and replaced with
VDATE. VDATE is the verification date to be done by the metp
tasks based on VRFYBACK_HRS set in config.metp.

Lines were added to catch the exit status of VERIFGLOBAL_SH
and to exit if not succesfully run so it can be caught by
rocoto.

Refs: #315