Add error checking to ww3prepost build
Added some rudimentary error checking to the ww3prepost build script.
Formerly any errors in building the programs was not reported by the
script by returning a non-zero error code, which caused build_all to
think it completed successfully.

Now, the return code from the compilation of each of the three programs
is checked to make sure it is non-zero, and the script will print an
error message and exit with the same error code if not.

Additionally, set -u is turned on to make sure that no undefined
variables are used. This required commenting out some currently unused
netcdf variables.

Refs: #339