Correct cycling interval issues for aerosols
FHCYC was being used to determine the forecast interval for the aerosol
cycling, but this appears to be an appropriate setting to use. Added a
new exported environment variable STEP_GFS, which is the number of hours
between GFS forecasts. While it is derived from the workflow STEP_GFS
setting, it is only the integer number of hours, not a timedelta like
the workflow setting. May want to revist this later to avoid confusion.

Additionally, the restart file written at the end of the forecast does
not have the timestamp as part of the filename like the intermediary
restart files do. The chem init script is adjusted to not use the time-
stamp when looking for the final restart file. And, since it now has to
read in FHMAX anyway, also added a short-circuit so the search doesn't
bother trying to search previous cycles that can't include the restart
time needed.

Refs: #314