Update CICE inputs for new model version
The newest version of CICE replaces istep0 with the normal month and day,
so the namelist is updated to reflect that.

ktherm and tfrz_option are updated to allow alternate settings, following
similar changes to the UFS test harness. The default values are the
previous settings.

Settings for tr_FY, sw_redist, f_mlt_onset, and f_frx_onset are updated
to new values in the UFS test harness.

Additionally, many namelist options have been removed following the same
deletions in the UFS test harness.

Refs: #324