Update UFS build script
Some changes to the UFS require updates to the build script.

The modulefile organization has been updated to be
modulefiles/ufs_${target} rather than modulefiles/${target}/fv3, so
the module paths in the build script are updated correspondingly.

A cmake switch DAPP is now required in order to build UFS. Non-
aerosol coupled is using S2SW for the time being. This replaces
setting DS2S and DWW3 to YES.

The old v15 suite is removed from the CCPP suite list.

Refs: #324