Update templates to set ocean mesh file
The latest version of UFS adds a setting to define the ocean mesh file,
so the nems.configure templates are updated to define them. They are
identical to the ice mesh file, so the MESHICE variable is renamed to
MESH_OCN_ICE and used for both. Additionally, the use_mommesh variable
is added and defaults to true.

Refs: #324