Update UFS to 20210720 develop
Updates the UFS version to the current develop.

The UFS build script in g-w is changed as the bulid script in the UFS
repo no longer automatically adds CCPP_FLAGS to CMAKE_FLAGS, so now the
g-w version does it instead before calling the UFS script. While this
change was being made, the option parsing was updated to use getopts to
match other scripts.

The FV3 namelist has ncld removed as that option is no longer recognized.

Refs: #324