Ensure CCPP_SUITE is set and suite file exists
Previously there was no guarantee that CCPP_SUITE was set before it
was used in forecast_postdet.sh. The suite how has a default value if
it is not set. A default later in the execution chain is removed, as
CCPP_SUITE is now guaranteed to be set beforehand. The default is also
updated to FV3_GFS_v16.

There is also now a check to ensure the suite file is present inside
the UFS directory before trying to determine settings from it.

Refs: #346