Update to MERRA2 aerosol climatology
Adds the capability to use MERRA2 aerosol climatology and makes it
the default.

An addition to the diag table is required. Rather than continue the
proliferation of diag tables to produce one specifically for coupled
and MERRA2, the existing diag table from the MERRA2 update in devleop
is pared down to just the additional fields introduced. This is then
appended to the main diag table if necessary.

Also corrected a related issue in the CROW forecast config. While
IAER and several other similar settings were already present in the
schema, the values were hard-coded in the forecast config instead of
using the value set. The script now correctly uses the values set by
the configuration system. The default for IAER is changed to 1011 for
MERRA2.

Refs: #379