Update namelist settings for prototype 7c
Several namelist settings are updated for prototype 7c.

A large number of the settings are tied to Noah-MP, so those settings
are now set alongside the lsm, but mostly accept overrides.

Another group of settings are related to cellular automata and are
only added to the namelist if a new setting, DO_CA, is set to YES.
The setting defaults to on (YES).

Two other settings, IALB and IEMS, are not confirmed to be tied to
other settings, so the defaults for these are just straightfowardly
changed to the new values.

Refs: #386