Choose nems.configure template correctly
The confignamevarfornems was not being set correctly, so the default
of 'atm' was always used. In coupled-crow, this variable was set to
nems_temp in the j-job because there is a separate j-job for cold-
starting the mediator. Here I just cut out the middle-man and changed
the nems_temp assignments to confignamevarfornems in config.base.

Refs: #416