Consolidate fcst resources and reorg configs
Resource settings for the forecast job were previously scattered
between multiple files. These settings are now consolidated in
config.resources.

The total number of PEs is now calculated properly if different
coupled models are included.

The config.fcst script is updated to properly source component
scripts if that component is on. New scripts for those components
were created where needed, and settings were moved into the
appropriate scripts (though it may be appropriate to relocate
additional settings in the future). The special load of the wave
config in the j-job is removed as it is no longer needed.

CDUMP assignment had to be moved up in the j-job to occur before
the config files are sourced so the correct settings are used.

Undefined DO switches for stochastic components are added to
prevent syntax errors downstream.

Settings for do_ugwp, do_tofd, and launch_level are removed from
the forecast config, as they are now set in the forecast script
based on the CCPP suite.

Refs: #416