Fix mosaic and mesh files
The wrong path to the mosaic fix file was being used for coupled
forecasts and the mesh file was not being determined correctly based
on the ice resolution. Both are now corrected. Since it is determined
by the ice resolution, the ocn/ice meshfile name is now determined in
the CICE post-det.

Refs: #416