Remove check for existence of tiled fix files
The existing of some fix files was being checked with fallback
values used if the files are missing. However, this approach does
not work with the tiled fix files, as they are passed as 'tileX',
while the files are tile1, tile2, etc. This was causing the files
checked to always fall back to their non-tiled versions.

In the future, we should check for the existence of all fix files,
not just the five that were checked here, appropriately handling
the tiled versions.

Refs: #416