Update GSI tag to gfsda.v16.1.3
The GSI tag was updated from gfsda.v16.1.1a to gfsda.v16.1.3.  This tag
turns off GeoOptics assimilation, turns on Spire assimilation, adjusts
the observation errors for the commerical data, corrects two RO
observation operator bugs, and relaxes the gross check for RO
that was added as a mitigation measure for those bugs.

Ref: #420