Fix coupling interval for atm-wav
The nems.configure template for atm-wave was still using the older
coupling_interval_sec, which was not being subsituted at runtime.
The parameter is updated to coupling_interval_slow_sec to match what
is used for the full coupled model, and the substitution is performed
whenever any coupling is on, instead of just cplflux.

Refs: #416