Stop double-parsing task list with rocotoviewer
The workflow was essentially being parsed a second time and only
classifying a task as a metatask if it has a certain naming conven-
tion, rather than obeying the actual XML specification. This change
was presumably made because CROW made every task part of a metatask
group, so this step was needed stop treating all tasks as metatasks.

This also brings up another issue: that nested metatasks aren't
currently handled by rocotoviewer. A new issue should be opened once
a permanent home for rocotoviwer is determined.