Update to UFS version as of 2021 Oct 07
This update adds several new stochastic physics possibilities, with
associated switched added. There are also some other new namelist
settings.

MOM input template files have been updated to use the @[ ] notation
following same changes in UFS, with corresponding changes to the
replacement code. The new template files were copied from UFS, except
saving of initial conditions is turned on.

Due to changes in how the output cadence is defined in the namelist,
an explicit list of forecast hours is generated in the pre-det script.
To avoid duplication, this list is used later to create the symlinks
for the output files.