Symlink ocean/wave output to component directories in COM
Ocean and ice output now have symlinks created before the forecast
is run so output is available in COM immediately instead of being
copied when the forecast is complete.

Ocean and ice output in COM are now also located in the appropriate
component directories like data for other components. For now, the
joint ocnice files are placed in the ocean directory. There are
ongoing discussions to determine how these will be handled in the
future.

Refs: #191, #483, #485