Add ability to turn off wave boundary point post jobs
Adds a setting, DOBNDPNT_WAVE, that allows the boundary point wave
post jobs to be turned off. The setting defaults to YES, but is NO
for the coupled defaults.

In order to make sure the wave config is included in the settings
dict for setup_workflow, workflow utils now includes the config.wave
file in the settings for all jobs that begin 'wave'. This allows the
retrieval of the DOBNDPNT_WAVE setting.

Additionally, the task definition for the gdas and gfs versions of
wavepostpnt is consolidated into a single block to eliminate the
redundant code.

Closes: #486