Add coupled tasks to forecast-only workflow setup
Tasks to stage coupled ICs and to run ocean post are added for the
forecast-only workflow setup.

The init jobs run is controlled by the app setting. If the app is
S2S or S2SW, the coupled IC job will run instead of getic and init.
The source locations for the coupled IC job are determined by the
settings in the config.coupled_ic file. There is a base location,
ORIGIN_ROOT, then the subdirectory for each component is given by a
different setting.

Ocean post is pooled into tasks in the same manner as atm post, with
the number of tasks specified in config.ocnpost.

The ocean post config file was missing a line to source the job
resources, which is needed for the workflow generation, so this was
updated.

Refs: #481, #482, #483