Only read from wavepostsbs when waves are on
Workflow generation would fail when waves weren't on because the
key 'wavepostsbs' wasn't defined when reading the whethern to run
wave boundary point jobs from the configs. Reading of other compon-
ent specific settings is not an issue because they are all behind
checks that the relevant component is on.

Refs: #486