Update gfsmetp and gfsarch dependencies
gfsmetp had a dependency on the previous cycle's
gfsarch, as gfsmetp uses the data moved to ARCDIR
by gfsarch. This caused problems in cycled runs where
the previous gfsarch cycle did not exist in the first
run of the gfs cycles. This removes that dependency.

A dependency gfsarch on gfsmetp was added. This was
added the running of the gfsmetp jobs will now use
data in ROTDIR, which gfsarch will possibly remove
as a part of clean up.

Refs: #437