Add coupled files to archive and streamline
The following tar files for ocean and ice output are added to the
archive job to be stored on HPSS:
- ocn_2D
- ocn_3D
- ocn_xsect
- ocn_daily
- ocn_ice_grib2_0p25
- ocn_ice_grib2_0p5
- wavocn
- ice
- gfs_flux_1p00

Handling of the archival of gaussian grid NetCDF files is changed. The
former, unused, and now misnamed setting of SAVEFCSTNEMSIO has been
changed to ARCH_GAUSSIAN and is now also used consistently between
arch.sh and hpssarch_gen.sh. gfs_pgrb2b remains controlled by this
setting as well, though it is not a gaussian grib.  Additionally,
there are new settings, ARCH_GAUSSIAN_FHMAX and ARCH_GAUSSIAN_FHINC,
to control which atm and sfc files are saved. The default values are
36 and 6, maintaining the previous hardcoded values

The archive job is also streamlined by replacing a bunch of repeating
code. Instead of creating tar files in small loops, a list of all the
tarfiles to archive is created, then a single loop executes the
archival and error checking. Only the MOS tarfile remains separate
(for now).

The forecast input.nml has been added to the gfsa tarball. The WW3
grid definitions (ww3_multi*) have been added to the gfswave tarball.

Refs: #416