Enable connectionless UDP for MPI
A new MPI setting to turn on connectionless UDP. This was found to
be necessary for machine stability when aerosols are turned on
(specifically for MAPL). Per "Optimizing HPC Applications with
Intel Cluster Tools: Hunting Petaflops", this setting should be on
when running with thousands of processes.

When we break the modulefiles apart by job, we might want to have
this change only apply to the forecast job.

Refs: #516