Add --aerosols option to setup expt
To control whether coupled aerosols are used, a new option is added
to the experiment setup scripts: --aerosols. It is an optional switch
and takes no positional arguments. If --aerosols is used, DO_AERO will
be set to "YES"; otherwise it will be "NO".

The ATMAERO app is removed as an option in config.base as redundant.
The setting had never been added to the setup expt scripts.

Refs: #516