Update diag table for coupled aerosols
Now that aerosols are permitted to run with coupled, a new diag table
is needed for that configuration. Instead of adding nested ifs to
handle the different combinations, the coupled diag table is moved
out into its own block, then if aerosols are on, '_aer' is appended to
the file name, whichever diag table is used.

In the future, I hope to break diag_table up into pieces that can be
assembled based on component (see Issue #567).

Refs: #516